pyproject.toml
[tool.poetry]
name = "findmyphone"
version = "0.1.0"
description = "Rest API to receive phones locations"
authors = ["Arthur <arp.arthur@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
python-dotenv = "^1.0.1"
SQLAlchemy = "^2.0.31"
django = "^5.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
setuptools = "^72.1.0"
wheel = "^0.44.0"
gdal = "^3.9.1"
psycopg2 = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true